refactor(commit): commit準備処理を分割しパッチ出力を追加
一時的にテキストエディタ向けのverbose表示は消えますが、
すぐに復活させる予定です。
